+++ src/main/webapp/META-INF/context.xml
@@ -5,7 +5,7 @@
               type="javax.sql.DataSource"
               username="root"
               password="coderslab"
-              driverClassName="com.mysql.jdbc.Driver"
+              driverClassName="com.mysql.cj.jdbc.Driver"
               connectionProperties="useUnicode=yes;characterEncoding=utf8;serverTimezone=UTC"
               url="jdbc:mysql://localhost:3306/workshop3"
               maxTotal="100"
