+++ src/main/webapp/header.jsp
@@ -46,7 +46,7 @@
 
         <!-- Nav Item - Dashboard -->
         <li class="nav-item active">
-            <a class="nav-link" href="list.jsp">
+            <a class="nav-link" href="<c:url value="/user/list" />">
                 <i class="fas fa-fw fa-tachometer-alt"></i>
                 <span>Dashboard</span></a>
         </li>
