+++ src/main/webapp/table.jsp
@@ -1,39 +1,32 @@
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<!-- DataTales Example -->
-<div class="card shadow mb-4">
-    <div class="card-header py-3">
-        <h6 class="m-0 font-weight-bold text-primary">Users List</h6>
-    </div>
-    <div class="card-body">
-        <div class="table-responsive">
-            <table class="table table-bordered" id="dataTable" width="100%" cellspacing="0">
-                <thead>
-                <tr>
-                    <th>ID</th>
-                    <th>Username</th>
-                    <th>Email</th>
-                    <th>Action</th>
-                </tr>
-                </thead>
-                <tbody>
-                <c:forEach items="${users}" var="user">
-                    <tr>
-                        <td><c:out value="${user.id}"/></td>
-                        <td><c:out value="${user.username}"/></td>
-                        <td><c:out value="${user.email}"/></td>
-                        <td class="text-left py-1 px-0">
-                            <div>
-                                <button type="button" class="btn btn-link">Delete</button>
-                                <button type="button" class="btn btn-link">Edit</button>
-                                <button type="button" class="btn btn-link">Show</button>
-                            </div>
-                        </td>
-                    </tr>
-                </c:forEach>
 
-                </tbody>
-            </table>
-        </div>
-    </div>
+<div class="table-responsive">
+    <table class="table table-bordered" id="dataTable" width="100%" cellspacing="0">
+        <thead>
+        <tr>
+            <th>ID</th>
+            <th>Username</th>
+            <th>Email</th>
+            <th>Action</th>
+        </tr>
+        </thead>
+        <tbody>
+        <c:forEach items="${users}" var="user">
+            <tr>
+                <td><c:out value="${user.id}"/></td>
+                <td><c:out value="${user.username}"/></td>
+                <td><c:out value="${user.email}"/></td>
+                <td class="text-left py-1 px-0">
+                    <div>
+                        <button type="button" class="btn btn-link">Delete</button>
+                        <button type="button" class="btn btn-link">Edit</button>
+                        <button type="button" class="btn btn-link">Show</button>
+                    </div>
+                </td>
+            </tr>
+        </c:forEach>
+
+        </tbody>
+    </table>
 </div>
